chore: wipe config
